BSDP: Allow "pad" and "end" options.
Apple bsdpd uses the same routine to parse BSDP options as it uses to parse the
DHCP options, which means that "pad" (0) and "end" (255) are accepted as BSDP
options: They do not have a length field and do not have any value, so they
always consume exactly one byte.  Interestingly, when bsdpd finds an "end" BSDP
option, it simply records that an "end" was encountered, and continues parsing.

This change enhances the BSDP option dissector to behave identically.

Testing Done: Loaded a BSDP ACK from a Dell KACE K2000 appliance; Previously it
	would warn about there being insufficient room for the length of the "end"
	option, and now it parses correctly.  Modified the packet to include a
	string of 0 and 255 options, and observed that they were parsed as expected:
	One byte each, no subtree, no length, and parsing continues afterwards.